api/requirements.txt
fastapi
uvicorn
Pillow
torch
torchvision
numpy
scipy